+++ src/use/store/mutation-types.ts
@@ -2,4 +2,5 @@ export enum MutationTypes {
     SET_LOGGEDIN = "SET_LOGGEDIN",
     SET_MODAL = "SET_MODAL",
     SET_USER = "SET_USER",
+    FORCE_CLOSE_BURGER_MENU = "FORCE_CLOSE_BURGER_MENU",
 }