+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - add profile picture functionality (set, update, delete) for admins and users [#551](https://github.com/upb-uc4/ui-web/pull/551)
 
 ## Bugfix
+- close mobile burger menu on navigation to another page [#552](https://github.com/upb-uc4/ui-web/issues/552)
 - remove stubs from the student's desktop navigation [#537](https://github.com/upb-uc4/ui-web/pull/537)
 - error feedback is given on posting an account without role [#535](https://github.com/upb-uc4/ui-web/issues/535)